Sample UDDL SDM added for test purposes
The complexity of UDDL means that parsing tests really need a lot just
to get started. This provides a foundation using a small version of the
UDDL SDM.

This way we don't embed this in code. Makes it easier to update if the
format changes.

Just read this file into an XTextResource and then add additional test
model in the same resource set.